# ArchitectureRefactoringHelper/Repository/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Repository, Version=1.0.0.0, Culture=neutral, PublicKeyToken=null",
    "version": "1.0"
  },
  "paths": {
...
    "/api/v1/processes/{id}": {
      "get": {
        "tags": [
          "ApproachProcess"
        ],
        "operationId": "GetApproachProcess",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApproachProcess"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/processes/qualities": {
      "get": {
        "tags": [
          "ApproachProcess"
        ],
        "operationId": "ListQualities",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Quality"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "ApproachProcess"
        ],
        "operationId": "AddQuality",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Quality"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Quality"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Quality"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Quality"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/processes/qualities/{name}": {
      "delete": {
        "tags": [
          "ApproachProcess"
        ],
        "operationId": "DeleteQuality",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/v1/processes/quality-sublevels": {
      "get": {
        "tags": [
          "ApproachProcess"
        ],
        "operationId": "ListQualitySublevels",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/QualitySublevel"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "ApproachProcess"
        ],
        "operationId": "AddQualitySublevel",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QualitySublevel"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/QualitySublevel"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/QualitySublevel"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QualitySublevel"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/processes/quality-sublevels/{name}": {
      "delete": {
        "tags": [
          "ApproachProcess"
        ],
        "operationId": "DeleteQualitySublevel",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/v1/processes/directions": {
      "get": {
        "tags": [
          "ApproachProcess"
        ],
        "operationId": "ListDirections",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Direction"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "ApproachProcess"
        ],
        "operationId": "AddDirection",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Direction"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Direction"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Direction"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Direction"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/processes/directions/{name}": {
      "delete": {
        "tags": [
          "ApproachProcess"
        ],
        "operationId": "DeleteDirection",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/v1/processes/automation-levels": {
      "get": {
        "tags": [
          "ApproachProcess"
        ],
        "operationId": "ListAutomationLevels",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AutomationLevel"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "ApproachProcess"
        ],
        "operationId": "AddAutomationLevel",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AutomationLevel"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/AutomationLevel"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/AutomationLevel"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AutomationLevel"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/processes/automation-levels/{name}": {
      "delete": {
        "tags": [
          "ApproachProcess"
        ],
        "operationId": "DeleteAutomationLevel",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/v1/processes/analysis-types": {
      "get": {
        "tags": [
          "ApproachProcess"
        ],
        "operationId": "ListAnalysisTypes",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AnalysisType"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "ApproachProcess"
        ],
        "operationId": "AddAnalysisType",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AnalysisType"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/AnalysisType"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/AnalysisType"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnalysisType"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/processes/analysis-types/{name}": {
      "delete": {
        "tags": [
          "ApproachProcess"
        ],
        "operationId": "DeleteAnalysisType",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/v1/processes/techniques": {
      "get": {
        "tags": [
          "ApproachProcess"
        ],
        "operationId": "ListTechniques",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Quality"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "ApproachProcess"
        ],
        "operationId": "AddTechnique",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Technique"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Technique"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Technique"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Technique"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/processes/techniques/{name}": {
      "delete": {
        "tags": [
          "ApproachProcess"
        ],
        "operationId": "DeleteTechnique",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/v1/processes/process-strategies": {
      "get": {
        "tags": [
          "ApproachProcess"
        ],
        "operationId": "ListProcessStrategies",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ProcessStrategy"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "ApproachProcess"
        ],
        "operationId": "AddProcessStrategy",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProcessStrategy"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ProcessStrategy"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/ProcessStrategy"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProcessStrategy"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/processes/process-strategies/{name}": {
      "delete": {
        "tags": [
          "ApproachProcess"
        ],
        "operationId": "DeleteProcessStrategy",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/v1/processes/atomar-units": {
      "get": {
        "tags": [
          "ApproachProcess"
        ],
        "operationId": "ListAtomarUnits",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AtomarUnit"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "ApproachProcess"
        ],
        "operationId": "AddAtomarUnit",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AtomarUnit"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/AtomarUnit"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/AtomarUnit"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AtomarUnit"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/processes/atomar-units/{name}": {
      "delete": {
        "tags": [
          "ApproachProcess"
        ],
        "operationId": "DeleteAtomarUnit",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AnalysisType": {
        "required": [
          "name"
        ],
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "ApproachProcess": {
        "type": "object",
        "properties": {
          "approachProcessId": {
            "type": "integer",
            "format": "int32"
          },
          "qualities": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Quality"
            },
            "nullable": true
          },
          "qualitySublevels": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/QualitySublevel"
            },
            "nullable": true
          },
          "directions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Direction"
            },
            "nullable": true
          },
          "automationLevels": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AutomationLevel"
            },
            "nullable": true
          },
          "analysisTypes": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AnalysisType"
            },
            "nullable": true
          },
          "techniques": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Technique"
            },
            "nullable": true
          },
          "processStrategies": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ProcessStrategy"
            },
            "nullable": true
          },
          "atomarUnits": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AtomarUnit"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "AtomarUnit": {
        "required": [
          "name"
        ],
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "AutomationLevel": {
        "required": [
          "name"
        ],
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "CalculationMetric": {
        "required": [
          "name"
        ],
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "Direction": {
        "required": [
          "name"
        ],
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "ProcessStrategy": {
        "required": [
          "name"
        ],
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "Quality": {
        "required": [
          "category",
          "name"
        ],
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "category": {
            "$ref": "#/components/schemas/QualityCategory"
          },
          "qualitySublevels": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/QualitySublevel"
            },
            "nullable": true
          },
          "qualityTradeOffs": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Quality"
            },
            "nullable": true
          },
          "calculationMetrics": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CalculationMetric"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "QualityCategory": {
        "enum": [
          "Attribute",
          "SystemProperty"
        ],
        "type": "string"
      },
...
      "QualitySublevel": {
        "required": [
          "name"
        ],
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "qualityName": {
            "type": "string",
            "nullable": true
          },
          "qualityTradeOffs": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Quality"
            },
            "nullable": true
          },
          "calculationMetrics": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CalculationMetric"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "Technique": {
        "required": [
          "name"
        ],
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
    }
  }
}